Create navigation-history.py
Navigation history is stored in Windows machines in the directory with the following path: C:\Users\(username)\AppData\Local\Google\Chrome\User Data\Default\History

We can check the contents of the database with DB Browser for SQLite. In the table we find the URLs of the websites visited, the title of the page, the number of visits and the last time the person visited the website.
This script extracts the information stored in the database and copy it to a .txt file "history" chronologically.